% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03datatrfsum.R
\name{ms_blank_sub}
\alias{ms_blank_sub}
\title{Blank Subtraction}
\usage{
ms_blank_sub(
  dat,
  dat_bl,
  md,
  md_var = "Group",
  sid = "SampleID",
  threshold = 3,
  sub_blanks = FALSE
)
}
\arguments{
\item{dat}{A data matrix containing compounds as rownames and sample IDs
as column names.}

\item{dat_bl}{A data matrix of method blank intensities for each compound
in the same format as the input data.}

\item{md}{A data frame containing metadata accompanying the data matrix.
Must have a column indicating sample groups.}

\item{md_var}{Grouping variable for data checks; this is generally
a variable containing the group IDs for each sample.}

\item{sid}{Sample ID variable.}

\item{threshold}{Signal intensity threshold for removing a compound from
the dataset. The default is to remove compounds that are < 3-fold the
signal intensity of the method blanks.}

\item{sub_blanks}{Should compounds present in blanks be removed?
If FALSE (default), only missing compounds are removed from the dataset.}
}
\value{
A filtered data frame retaining compounds that exceed the
blank intensity threshold.
}
\description{
Removes compounds from a dataset based on intensity in method blanks.
}
\examples{

# ms_blank_sub(
#   dat = ld[["wash"]][["input"]],
#   dat_bl = ld[["qc"]][["input"]],
#   md = ld[["wash"]][["meta"]]
# )

}
