% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/031QCreport.r
\name{ms_check_pres}
\alias{ms_check_pres}
\title{QC Check Group Presence}
\usage{
ms_check_pres(
  exp1,
  asy = "raw",
  var_g = "Group",
  var_feat = "Label",
  hw = 20,
  hh = 20,
  fsc = 8,
  fsr = 8,
  clc = TRUE,
  clr = TRUE,
  showcols = TRUE,
  showrows = TRUE,
  trans = TRUE,
  title1 = NULL
)
}
\arguments{
\item{exp1}{A SummarizedExperiment.}

\item{asy}{Assay name to calculate group presence/absence.}

\item{var_g}{Group variable name.}

\item{var_feat}{Feature name column from rowData.}

\item{hw}{Presence/absence heatmap width.}

\item{hh}{Presence/absence heatmap height.}

\item{fsc}{Heatmap column names fontsize.}

\item{fsr}{Heatmap row names fontsize.}

\item{clc}{Cluster heatmap columns?}

\item{clr}{Cluster heatmap rows?}

\item{showcols}{Show heatmap column names?}

\item{showrows}{Show heatmap row names?}

\item{trans}{Transpose heatmap?}

\item{title1}{Heatmap title.}
}
\value{
Heatmap displaying compound presence/absence.
}
\description{
Checks compound presence per treatment group for the selected dataset
and flags compounds that are not present in at least 50% of samples
in one group.
}
\examples{

# ms_check_pres(data1)


}
