% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01params.R
\name{ms_create_exp}
\alias{ms_create_exp}
\title{Create SummarizedExperiment from Excel}
\usage{
ms_create_exp(
  file_data,
  file_md = NULL,
  sheetno = 1,
  md_no_feat,
  md_no_samp,
  format_data = TRUE
)
}
\arguments{
\item{file_data}{Path to input Excel file.}

\item{file_md}{(Optional) Path to input Excel file containing study metadata.}

\item{sheetno}{Numerical value indicating the sheet number containing
input data.}

\item{md_no_feat}{Number of compound/feature metadata columns.}

\item{md_no_samp}{Number of sample metadata columns.}

\item{format_data}{If TRUE, format input data matrix to remove NA values
and correct data type.}
}
\value{
A SummarizedExperiment containing the data matrix
with the final feature and sample metadata
for the specified ionization mode(s).
}
\description{
Creates a SummarizedExperiment for use with COMb functions from an
Excel file. This function produces an output equivalent to ms_postproc,
which is useful in cases where processed data is already available
for analysis. The input file should contain 3 essential components:
the data matrix, compound/feature metadata, and sample metadata.
Optionally, the user may provide an additional file containing study
metadata for use with specific COMb functions, including ms_qc and
ms_qc_report.
}
\examples{

# data1 <- ms_create_exp(
#   "path/to/excelfile.xlsx",
#   md_no_feat = 10,
#   md_no_samp = 5
# )

}
