% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04datatrfsum.R
\name{ms_data_check}
\alias{ms_data_check}
\title{Data Transformation, Scaling and Quality Check}
\usage{
ms_data_check(
  dat,
  md,
  md_var = "Group",
  sid = "Label",
  sc_meth = "Pareto",
  dimtype = "UMAP",
  plab = TRUE,
  hw = 20,
  hh = 20,
  fsr = 8,
  fsc = 8,
  trans = FALSE
)
}
\arguments{
\item{dat}{A data matrix containing compounds as rownames and sample IDs
as column names.}

\item{md}{A data frame containing metadata accompanying the data matrix.}

\item{md_var}{Grouping variable for data checks; this is generally
a variable containing the group IDs for each sample.}

\item{sid}{Sample ID variable.}

\item{sc_meth}{Data scaling method (performed after log2 transformation);
Either "Median" or "Pareto" (default).}

\item{dimtype}{Dimension reduction to calculate for scaled data; either "PCA"
or "UMAP" (default).}

\item{plab}{Should plot labels be shown on the loadings plot?}

\item{hw}{Correlation heatmap width.}

\item{hh}{Correlation heatmap height.}

\item{fsr}{Heatmap row names fontsize.}

\item{fsc}{Heatmap column names fontsize.}

\item{trans}{Transpose input data?}
}
\value{
List containing formatted input data and metadata
for downstream analysis.
}
\description{
Performs data imputation, log2-transformation, scaling, and
quality checks on an untargeted dataset.
}
\examples{

# ms_data_check(d, md)

}
