% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03datanorm.r
\name{ms_data_norm}
\alias{ms_data_norm}
\title{Data Normalization}
\usage{
ms_data_norm(ld1, mtd = "iSTD")
}
\arguments{
\item{ld1}{A SummarizedExperiment object, usually generated by ms_postproc.}

\item{mtd}{Normalization method (either "tic", "iSTD", or "LOESS").}
}
\value{
A SummarizedExperiment containing normalized
compound intensities and metadata.
}
\description{
Includes multiple methods for data normalization. Currently supported methods
are internal standard. Planned methods include total ion count (TIC), LOESS,
and SERRF.
}
\examples{

# data1 <- ms_data_norm(ld1 = data1)

}
