% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07dimrd.R
\name{ms_dim_rd}
\alias{ms_dim_rd}
\title{Dimension Reduction}
\usage{
ms_dim_rd(
  exp1,
  asy = "norm.pareto",
  md_var,
  dim_type = "PCA",
  ret_umap = FALSE,
  dim1 = "2D",
  p_lab = TRUE,
  show_axes = TRUE,
  namefile = NULL,
  flag_outliers = FALSE,
  outlier_calc = NULL,
  sid = NULL,
  scale_data = TRUE
)
}
\arguments{
\item{exp1}{SummarizedExperiment object.}

\item{asy}{Assay to use for dimension reduction.}

\item{md_var}{Name of column data variable for plot overlay.}

\item{dim_type}{Dimension reduction technique to use for plotting.
Currently available options are "PCA", "UMAP", or "PLSDA".}

\item{ret_umap}{Only used if dim_type = "UMAP"; Use Python umap-learn
for calculating UMAP (TRUE or FALSE).}

\item{dim1}{Number of dimensions to plot data (Either "2D" or "3D").
Only 2D plots are used if plotting multiple panels.}

\item{p_lab}{Label groups on plot panels (Only available if "2D").}

\item{show_axes}{Should the x and y axis information be shown on the plot?}

\item{namefile}{If dim1 is "3D", provide the file name for saving the plot.}

\item{flag_outliers}{If TRUE, flag samples that are +/- 3 standard
deviations from the individual group means. Generally used only when
evaluating results from ms_data_norm.}

\item{outlier_calc}{If flag_outliers is TRUE, provide the column name
containing the TIC for determining outliers.}

\item{sid}{If flag_outliers is TRUE, provide the name of the sample ID
column for labeling outliers.}

\item{scale_data}{If TRUE, data will be pareto scaled
prior to dimension reduction.}
}
\value{
A dimension reduction plot of the specified type.
}
\description{
Performs dimension reduction for a selected dataset using one
of the following methods: PCA, UMAP, PLSDA.
}
\examples{

# ms_dim_rd(
#   exp1 = data1,
#   md_var = "Group"
# )

}
