% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/031QCreport.r
\name{ms_group_rsd}
\alias{ms_group_rsd}
\title{Normalization QC RSD}
\usage{
ms_group_rsd(exp1, asy = "imputed", var_g = "Group", var_feat = "Label")
}
\arguments{
\item{exp1}{SummarizedExperiment containing sample metadata.}

\item{asy}{Assay for calculating RSD values.}

\item{var_g}{Grouping variable.}

\item{var_feat}{Feature name column from rowData.}
}
\value{
Group RSD values for each feature.
}
\description{
Calculates relative standard deviation for each feature per
group within a SummarizedExperiment object.
}
\examples{

# ptic <- ms_group_rsd(
#   exp1 = data1
# )

}
