% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10visualize.R
\name{ms_plot_bar}
\alias{ms_plot_bar}
\title{Bar Plot}
\usage{
ms_plot_bar(
  ptype = "stacked",
  df1,
  df2 = NULL,
  xvar = "Assay",
  yvar = "Count",
  clvar = "Class",
  labvar = NULL,
  plab = "Proportion"
)
}
\arguments{
\item{ptype}{Bar plot type (either "stacked" or "std").}

\item{df1}{Input data frame.}

\item{df2}{Data frame containing group standard deviations (for error bars).}

\item{xvar}{X-axis variable.}

\item{yvar}{Y-axis variable.}

\item{clvar}{Group variable.}

\item{labvar}{Additional variable for plotting proportions (optional).}

\item{plab}{Plot y-axis label.}
}
\value{
A bar plot for the chosen treatment comparisons.
}
\description{
Generates a bar plot with error bars from an input data frame.
}
\examples{

# ms_plot_bar(
#   df1 = dox,
#   xvar = "rt",
#   yvar = "mz"
# )

}
