% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10visualize.R
\name{ms_plot_heat}
\alias{ms_plot_heat}
\title{Heatmap 1}
\usage{
ms_plot_heat(
  d_stat,
  d_ref = NULL,
  c_list,
  comp_col = "Comparison.fc",
  diff_col = "Log2FC",
  p_col = "p.value",
  an1 = NULL,
  hm_w = 36,
  hm_h = 10,
  fs_r = 8,
  fs_c = 8,
  cl_c = TRUE,
  cl_r = TRUE,
  cn = TRUE,
  rn = TRUE,
  filtp = FALSE,
  scm1 = col_grad(scm = 3),
  col_fun = 1
)
}
\arguments{
\item{d_stat}{Univariate results list generated by ms_stat_anova().}

\item{d_ref}{Reference annotation list generated by ms_input(),
which is only used if plotting a heatmap annotation.}

\item{c_list}{Comparison list, provided as a vector.}

\item{comp_col}{Fold change comparison column.}

\item{diff_col}{Column name containing effect size.}

\item{p_col}{Column name containing adjusted p-values.}

\item{an1}{Variable names used for annotating the heatmap.}

\item{hm_w}{Heatmap width.}

\item{hm_h}{Heatmap height.}

\item{fs_r}{Row fontsize.}

\item{fs_c}{Column fontsize.}

\item{cl_c}{Logical indicating if the columns should be clustered.}

\item{cl_r}{Logical indicating if the rows should be clustered.}

\item{cn}{Show column names.}

\item{rn}{Show row names.}

\item{filtp}{Filter heatmap based on a p-value cutoff of 0.05.}

\item{scm1}{Heatmap color scheme to use.}

\item{col_fun}{Color ramp function
(either by 4 quantiles [1] or by upper and lower quantiles[2]).}
}
\value{
A heatmap for the selected comparisons.
}
\description{
Generates a heatmap from a statistical results object for
a vector of comparisons.
}
\examples{

# ms_plot_heat(
#   d_stat = d[["stats"]],
#   c_list = c("KO vs. Control")
# )

}
