% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10visualize.R
\name{ms_plot_hm}
\alias{ms_plot_hm}
\title{General Heatmap}
\usage{
ms_plot_hm(
  so = NULL,
  stat_data = NULL,
  deg_sig = FALSE,
  l_cstm = NULL,
  asy = "scaled",
  asy_type = "scaled",
  cl_var = "Group",
  feat_var = "Label",
  h_w = 40,
  h_h = 20,
  fs_c = 6,
  fs_r = 8,
  cl_c = FALSE,
  cl_r = FALSE,
  rot_c = 45,
  rc = TRUE,
  rn = TRUE,
  col1 = col_grad(scm = 5),
  transp = FALSE,
  legend1 = "Abundance"
)
}
\arguments{
\item{so}{A SummarizedExperiment object.}

\item{stat_data}{Generate heatmap based on statistical results instead
of intensity/abundance values from a summarizedexperiment. Results
tables should be formatted according to outputs from ms_stat_uni.}

\item{deg_sig}{Should only compounds/features with statistically significant
differences in intensity/abundance be plotted?}

\item{l_cstm}{(optional) Filter data based on custom compound list.}

\item{asy}{Assay to use (ex. "norm" or "scaled").}

\item{asy_type}{Assay type (either "norm" or "scaled"). If "scaled",
the input data undergo no additional scaling.}

\item{cl_var}{Grouping variable name.}

\item{feat_var}{Variable from rowData containing compound names.}

\item{h_w}{Numeric value for heatmap width (passed to ComplexHeatmap).}

\item{h_h}{Numeric value for heatmap height (passed to ComplexHeatmap).}

\item{fs_c}{Numeric value for column fontsize (passed to ComplexHeatmap).}

\item{fs_r}{Numeric value for row fontsize (passed to ComplexHeatmap).}

\item{cl_c}{Cluster columns?}

\item{cl_r}{Cluster rows?}

\item{rot_c}{Rotation of column names.}

\item{rc}{Show column names?}

\item{rn}{Show row names?}

\item{col1}{Gradient color scheme to use.}

\item{transp}{Transpose heatmap input matrix (FALSE by default).}

\item{legend1}{Figure legend name.}
}
\value{
A ComplexHeatmap object.
}
\description{
Generates a heatmap from a Seurat Object,
which can be filtered based on differentially
expressed genes.
}
\examples{

# ms_plot_hm(so = data1)

}
