% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09visualize.R
\name{ms_plot_scatter}
\alias{ms_plot_scatter}
\title{Scatter Plot}
\usage{
ms_plot_scatter(df1, type1, xvar, yvar, clvar, oxvar = "ox", bns = 50)
}
\arguments{
\item{df1}{Input data frame.}

\item{type1}{Scatter plot type (either "std" or "oxview").
If type is "oxview," a variable indicating either "Oxidized"
or "Unoxidized" lipids must be present in the input data frame.}

\item{xvar}{X-axis variable.}

\item{yvar}{Y-axis variable.}

\item{clvar}{Group variable.}

\item{oxvar}{(optional) If type is "oxview," the variable name
distinguishing oxidized from unoxidized lipids.}

\item{bns}{(optional) Bin number for scatter plot types using
stat_density_2d().}
}
\value{
A scatter plot for the chosen treatment comparisons.
}
\description{
Generates a scatter plot from a statistical results object for
a chosen comparison.
}
\examples{

# ms_plot_scatter(
#   df1 = dox,
#   type1 = "oxview",
#   xvar = "rt",
#   yvar = "mz",
#   clvar = "class"
# )

}
