% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/031QCreport.r
\name{ms_plot_tic}
\alias{ms_plot_tic}
\title{Normalization QC Scatter Plot}
\usage{
ms_plot_tic(exp1, var_x, var_g = "Group", tic1, tic2)
}
\arguments{
\item{exp1}{SummarizedExperiment containing sample metadata.}

\item{var_x}{X-axis variable (usually sample ID).}

\item{var_g}{Grouping variable.}

\item{tic1}{Pre-normalization TIC column.}

\item{tic2}{Post-normalization TIC column.}
}
\value{
A scatter plot visualizing overall sample intensities.
}
\description{
Scatter plot visualization to evaluate normalization performance.
}
\examples{

# ptic <- ms_plot_tic(
#   exp1 = data1,
#   var_x = "sampleID",
#   var_g = "Group",
#   tic1 = "TIC.imputed",
#   tic2 = "TIC.norm"
# )

}
