% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10visualize.R
\name{ms_plot_vio}
\alias{ms_plot_vio}
\title{Violin/Box Plot}
\usage{
ms_plot_vio(
  exp1,
  asy = "norm",
  ptype = "vio",
  var_x = "Region",
  var_c = "Label",
  var_cl = "Group",
  var_sum = NULL,
  var_id = NULL,
  comp,
  plab = "Relative Abundance",
  title1 = NULL,
  show_jitter = FALSE
)
}
\arguments{
\item{exp1}{SummarizedExperiment object.}

\item{asy}{Assay to use for plotting.}

\item{ptype}{Plot type (either "vio" or "box")}

\item{var_x}{X-axis variable name from colData.}

\item{var_c}{Variable containing compound names.}

\item{var_cl}{Group variable name from colData.}

\item{var_sum}{Variable name for summarizing
compound information (optional).}

\item{var_id}{Sample ID variable name; only
required if var_sum is provided.}

\item{comp}{Compound name to plot.}

\item{plab}{Y-axis label.}

\item{title1}{Plot title.}

\item{show_jitter}{Show individual data points on the plot?}
}
\value{
A violin plot for the chosen treatment comparisons.
}
\description{
Generates a violin plot from an input data frame.
}
\examples{

# ms_plot_vio(
#   exp1 = data2,
#   comp = "Compound1"
# )

}
