% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02dataproc.r
\name{ms_postproc}
\alias{ms_postproc}
\title{Post-processing of Curated Lipidomics/Metabolomics Data}
\usage{
ms_postproc(
  dat,
  dat_in,
  col_id = "Alignment.ID.original",
  col_ad = "Alignment.ID.matched",
  samp_id = "ID",
  add_metadata = NULL
)
}
\arguments{
\item{dat}{Curated feature list, which is formatted as either a .csv
or tab-delimited .txt file. A column named "Mode" indicating the
ionization mode should be present in the file.}

\item{dat_in}{A RDS file generated by ms_preproc containing peak intensity
and sample information. If two file names are provided (e.g. pos and neg
ionization modes), then the data are merged into a single
SummarizedExperiment.}

\item{col_id}{The name of the column containing the original alignment IDs
for each feature in an exported MS-DIAL feature list. MUST match alignment
IDs in the peak intensity matrix to function properly.}

\item{col_ad}{Alignment ID column name; combines all features with duplicate
IDs, assuming that duplicate IDs correspond to adducts of the same feature.}

\item{samp_id}{Sample ID column name.}

\item{add_metadata}{Add additional metadata information to the output object.
These data should be provided as named list elements and are useful for
simplifying the use of COMb functions. See examples of common metadata below.}
}
\value{
A SummarizedExperiment containing the data matrix
with the final feature and sample metadata for the specified ionization mode.
}
\description{
Combines flagged adducts in a curated feature list by alignment
ID and filters input data for downstream normalization and analysis.
The curated feature list should be free of duplicate compounds
across both ionization modes (if applicable).
}
\examples{

# postproc <- ms_postproc(
#   dat = "data/03postprocessing/Chase_LSAE_curated.csv",
#   ionmode = "pos",
#   dat_in = "data/02preprocessing/Chase_LSAE_preprocessed_pos.rds"
# )

}
