% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02dataproc.r
\name{ms_preproc}
\alias{ms_preproc}
\title{Pre-processing of Exported MS-DIAL Alignment Results}
\usage{
ms_preproc(
  dattype = "lipidomics",
  ionmode = "pos",
  dir1 = "data/01alignment/",
  dat,
  ref1,
  ref2 = NULL,
  tolmz = 0.01,
  tolmz2 = 0.01,
  tolrt = 0.15,
  tolrt2 = 0.01,
  gcol = "Group",
  gblnk = "Blank",
  blank_filt = 5,
  min_fill = 0.01
)
}
\arguments{
\item{dattype}{Data type (either "lipidomics" or "metabolomics"). If the data
type is "lipidomics," features will be flagged for the presence of
characteristic MS/MS fragments for specific lipid classes.}

\item{ionmode}{Ionization mode (either "pos" or "neg").}

\item{dir1}{Directory containing the exported alignment file.}

\item{dat}{File name of an alignment result directly exported from MS-DIAL.}

\item{ref1}{Name of a text file containing sample metadata. Must ensure that
the order of samples listed in the alignment file matches the order of the
input sample list.}

\item{ref2}{(optional) The path to a text file containing compound names,
m/z, and retention time information for annotating unknowns.}

\item{tolmz}{m/z tolerance window (in mDa) for annotating unknown peaks
from a reference mzrt list.}

\item{tolmz2}{m/z tolerance window (in min) for flagging adducts and isotopes
of the same peak.}

\item{tolrt}{RT tolerance window (in min) for annotating unknown peaks
from a reference mzrt list.}

\item{tolrt2}{RT tolerance window (in min) for flagging adducts and isotopes
of the same peak.}

\item{gcol}{Sample group column name.}

\item{gblnk}{Method blank group name for calculating sample/blank intensity
ratio.}

\item{blank_filt}{Numerical value indicating the threshold for removing
compounds based on intensity in method blanks.
Defaults to 5x sample/blank average intensity per group.}

\item{min_fill}{Minimum fill percentage (expressed as a proportion) for
including features in the preprocessed output.}
}
\value{
A chromatogram based on X and Y coordinates.
}
\description{
Performs processing of alignment results exported from MS-DIAL
(v5.5 and higher). Adds a column flagging features for removal
based on presence in blanks, low abundance, or duplicate IDs.
Optionally annotates unknown features from matching to a reference
mzrt list.
}
\examples{

# preproc <- ms_preproc(
#   dat = "20250919_Chase_LSAE_CSH_pos_alignment",
#   ref1 = "ref/sample_list.csv",
#   ref2 = "ref/pos_mzrt_list.txt"
# )

}
