% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/031QCreport.r
\name{ms_qc}
\alias{ms_qc}
\title{COMb QC Report}
\usage{
ms_qc(exp1)
}
\arguments{
\item{exp1}{SummarizedExperiment containing one or more assays for
evaluating QC metrics.}
}
\value{
A markdown report containing all quality control information.
}
\description{
Performs quality control for a SummarizedExperiment that can be
summarized as a markdown report of key quality control metrics,
to assess normalization performance, including: dataset information,
data distribution, technical variance, sample correlation,
and outlier detection. The following minimum information should be
specified in the object metadata and included in the column
and/or feature data:
'Group column name', 'Acquisition order column name',
'Sample column name', 'QC sample name', 'Sample ID column name'.
Additionally, include 'Blank sample name' if the data contain
analytical method blank samples and 'Batch column name' if the
data are intended to be analyzed as separate batches
(ex. two distinct analytical runs or two separate
sample types, regions, matrices, etc.). By specifying the
'Batch column name', this function will automatically split the
data by batch and ionization mode. If 'Batch column name' is not
specified, the function will assume all data belong to the same
analytical run and will only be split by ionization mode.
}
\examples{

# ms_qc(data1)

}
