% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/031QCreport.r
\name{ms_qc_report}
\alias{ms_qc_report}
\title{QC Summary Report}
\usage{
ms_qc_report(
  data_qc,
  study_name = NULL,
  author = NULL,
  an_plat = NULL,
  an_date = NULL,
  meth_dp = NULL,
  meth_nm = NULL,
  outdir,
  outname
)
}
\arguments{
\item{data_qc}{Input object from ms_qc.}

\item{study_name}{Study name.}

\item{author}{Report author name.}

\item{an_plat}{Name of analysis platform.}

\item{an_date}{Analysis date(s).}

\item{meth_dp}{Data processing software/methods.}

\item{meth_nm}{Normalization method.}

\item{outdir}{Output directory.}

\item{outname}{Report filename.}
}
\value{
Group RSD values for each feature.
}
\description{
Creates a markdown-style report from a ms_qc QC summary.
}
\examples{

# ms_qc_report(
#   qcdata,
#   outdir = "path/to/directory/",
#   outname = "qcreport"
# )

}
