% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/031QCreport.r
\name{ms_samp_cor}
\alias{ms_samp_cor}
\title{Normalization QC Sample Correlation}
\usage{
ms_samp_cor(
  exp1,
  asy = "norm",
  var_g = "Group",
  sid = "Label",
  hh = 18,
  fsr = 6,
  pname = NULL
)
}
\arguments{
\item{exp1}{SummarizedExperiment containing sample metadata.}

\item{asy}{Assay for calculating RSD values.}

\item{var_g}{Grouping variable.}

\item{sid}{Sample ID column name.}

\item{hh}{Heatmap height/width.}

\item{fsr}{Row/column font size.}

\item{pname}{Plot name.}
}
\value{
Sample correlation heatmap.
}
\description{
Calculates correlations between study samples
within a SummarizedExperiment object.
}
\examples{

# ms_samp_cor(
#   exp1 = data1
# )

}
