% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11correlation.R
\name{ms_stat_corr}
\alias{ms_stat_corr}
\title{Correlation Analysis}
\usage{
ms_stat_corr(
  ptitle = "Correlation Matrix",
  exp1,
  asy1 = "norm",
  exp2 = NULL,
  asy2 = "norm",
  var_cl1 = NULL,
  var_cl2 = NULL,
  mode_cor = "features",
  filt_sig = TRUE,
  snc = TRUE,
  snr = TRUE,
  hh = 20,
  hw = 20,
  fsc = 8,
  fsr = 8,
  hmlinewidth = 0
)
}
\arguments{
\item{ptitle}{Heatmap title.}

\item{exp1}{SummarizedExperiment object.}

\item{asy1}{Assay to use for the first matrix.}

\item{exp2}{(optional) Second SummarizedExperiment object or
matrix if conducting correlation analysis between two matrices.}

\item{asy2}{Assay to use for the second matrix, if applicable.}

\item{var_cl1}{Class or sample group variable for adding an
annotation to the y-axis (exp1) of the correlation heatmap.}

\item{var_cl2}{Class or sample group variable for adding an
annotation to the x-axis (exp2) of the correlation heatmap.}

\item{mode_cor}{Should correlation coefficients be calculated
based on "samples" (columns) or "features" (rows)?}

\item{filt_sig}{Show only significant correlations on the heatmap?}

\item{snc}{Show heatmap column names?}

\item{snr}{Show heatmap row names?}

\item{hh}{Heatmap height.}

\item{hw}{Heatmap width.}

\item{fsc}{Heatmap column font size.}

\item{fsr}{Heatmap row font size.}

\item{hmlinewidth}{Heatmap box line width.}
}
\value{
A correlation analysis result list.
}
\description{
Conducts correlation analysis for each comparison in a
provided SummarizedExperiment. Returns the top compounds for
each correlation pair, statistical results, and a correlation
heatmap. Also supports inputs of two different matrices for
calculating pairwise correlations.
}
\examples{
# ms_stat_corr(
#   exp1 = d1,
#   asy1 = "raw"
# )

}
