% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09chemrich.R
\name{ms_stat_crich}
\alias{ms_stat_crich}
\title{ChemRICH Analysis Function}
\usage{
ms_stat_crich(
  d_stat,
  d_ref,
  col_class = "Class",
  col_lab = "Label",
  col_pvalue = "p adj"
)
}
\arguments{
\item{d_stat}{Univariate results list generated by ms_stat_uni.
Note that only "standard" ms_stat_uni inputs are supported as
"class" and "HiVE" inputs already contain results aggregated by class.}

\item{d_ref}{Data frame to use for assigning annotation classes.}

\item{col_class}{Variable name to use for assigning classes and grouping
for enrichment analysis.}

\item{col_lab}{Compound name column.}

\item{col_pvalue}{P-value column name in the input statistical results.}
}
\value{
A ChemRICH results data frame.
}
\description{
Conducts ChemRICH analysis for each comparison in a provided result
generated by ms_stat_uni(). The original enrichment method is adapted
from Chemical Similarity Enrichment Analysis (ChemRICH), which is detailed
in Barupal et. al. 2017 (DOI: 10.1038/s41598-017-15231-w).
Significantly enriched classes are determined by a weighted
Kolmogorov-Smirnov Test adjusted for false discovery rate (FDR).
}
\examples{
# ms_stat_crich(
#   d_stat = data2,
#   d_ref = df1
# )

}
