% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06fc.R
\name{ms_stat_fc}
\alias{ms_stat_fc}
\title{Group-wise Fold Change Function}
\usage{
ms_stat_fc(
  exp1,
  asy = "norm",
  mat_type = "norm",
  col_grp = "Group",
  col_class = "Class",
  col_lab = "Label",
  fc_method = "standard",
  comp_grp = TRUE
)
}
\arguments{
\item{exp1}{A SummarizedExperiment.}

\item{asy}{Assay to use for fold change calculation.}

\item{mat_type}{Input data type, either "norm" or "scaled." If
"scaled," fold changes are calculated as the difference between
two group averages rather than by dividing normalized compound
average intensities per group.}

\item{col_grp}{Vector of metadata variables for splitting data.}

\item{col_class}{Compound class variable name; only required if
fc_method is "class" or "HiVE".}

\item{col_lab}{Compound name column.}

\item{fc_method}{Fold change calculation mode; either "standard",
which is the most common, "class", or "HiVE"}

\item{comp_grp}{Only applicable if fc_method is "HiVE"; Should
group fold changes be calculated? If FALSE, returns ratios
without aggregation by treatment group.}
}
\value{
A data frame containing the group-wise
fold changes for each compound or class.
}
\description{
Calculates the log2 fold changes for all
possible combinations of selected metadata variables.
}
\examples{

# ms_stat_fc(exp1 = data2)

}
