% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08statsunivar.R
\name{ms_stat_uni}
\alias{ms_stat_uni}
\title{Multivariate ANOVA}
\usage{
ms_stat_uni(
  exp1,
  asy = "norm",
  mat_type = "norm",
  col_grp = "Group",
  col_class = "Class",
  col_lab = "Label",
  fc_method = "standard",
  name_group = NULL
)
}
\arguments{
\item{exp1}{A SummarizedExperiment object.}

\item{asy}{Input assay for statistical analysis.}

\item{mat_type}{Input data type, either "norm" or "scaled." If
"scaled," fold changes are calculated as the difference between
two group averages rather than by dividing normalized compound
average intensities per group.}

\item{col_grp}{Grouping variables for testing.}

\item{col_class}{Compound class variable name; only required if
fc_method is "class" or "HiVE".}

\item{col_lab}{Compound name column.}

\item{fc_method}{Fold change calculation mode; either "standard",
which is the most common, "class", or "HiVE".}

\item{name_group}{If statistics comparing two groups is desired,
provide a vector containing each group name for comparison.}
}
\value{
A data frame containing the combined ANOVA and fold change results
for the specified variable(s).
}
\description{
Conducts an ANOVA for each metadata variable with Tukey's post-hoc
analysis for each compound and combines with fold changes
for each group comparison.
}
\examples{

# ms_stat_uni(exp1 = data2)

}
