% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07statsunivar.R
\name{ms_stat_univ}
\alias{ms_stat_univ}
\title{T-test/Wilcox Rank Sum Test}
\usage{
ms_stat_univ(
  matpv,
  matfc,
  matfc_type = "norm",
  md,
  md_var = "Group",
  an = NULL,
  grp1,
  grp2,
  sid = "SampleID"
)
}
\arguments{
\item{matpv}{Data matrix for conducting ANOVA.}

\item{matfc}{Data matrix for calculating fold change.}

\item{matfc_type}{Input data type for calculating fold change;
either "norm" or "scaled."}

\item{md}{Metadata from ms_input().}

\item{md_var}{Metadata variables to test by ANOVA.}

\item{an}{Reference annotation data frame.}

\item{grp1}{Treatment group "A" name for test input.}

\item{grp2}{Treatment group "B" name.}

\item{sid}{Sample ID column for filtering input data matrix.}
}
\value{
A data frame containing the combined stats and fold change results
for the specified variable.
}
\description{
Conducts either Student's T-test or Wilcox Rank Sum Test
for selected variable, depending on data distribution,
and combines with fold changes for the selected comparison.
}
\examples{

# ms_stat_univ(
#   matpv = lstat[["data"]],
#   matfc = lstat[["data"]],
#   md = lstat[["meta"]],
#   grp1 = "Group1",
#   grp2 = "Group2"
# )

}
