% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00lib.R
\name{ms_theme_alt1}
\alias{ms_theme_alt1}
\title{Alternate Plot Theme}
\usage{
ms_theme_alt1(leg = c(0.95, 0.95))
}
\arguments{
\item{leg}{Plot legend coordinates.}
}
\value{
ggplot2 theme parameters to replace default plot theme.
}
\description{
Alternate plotting theme showing more details on the x and y axes.
}
\examples{

# ms_theme_alt1()

}
