% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02dataproc.r
\name{tms_chromatogram}
\alias{tms_chromatogram}
\title{Chromatogram Viewer (Targeted)}
\usage{
tms_chromatogram(
  dat,
  intype = "Skyline",
  comp,
  sid,
  sidcol = "SampleID",
  xcol = "RT_all",
  ycol = "Intensity",
  rtcol = "pkRT",
  lab = "SampleID",
  pkview = TRUE,
  pkcol = "pkBoundary"
)
}
\arguments{
\item{dat}{Input data list.}

\item{intype}{Input data type: either "Skyline," a list of 2 vectors
containing values separated by commas (default) or "xy," which is
a list of two variables containing vectors of x and y coordinates.}

\item{comp}{Compound name.}

\item{sid}{Sample ID.}

\item{sidcol}{Sample ID column name.}

\item{xcol}{X column name.}

\item{ycol}{Y column name.}

\item{rtcol}{Retention time column containing the specific retention
times of identified peaks.}

\item{lab}{Column to use for labeling samples.}

\item{pkview}{Highlight detected peaks if present in input data.
Requires result from tms_peakdetect.}

\item{pkcol}{Column generated from tms_peakdetect providing the
upper and lower RT limits of the detected peak.}
}
\value{
A chromatogram based on X and Y coordinates.
}
\description{
Plots chromatograms for selected samples and compounds from
input x and y coordinates.
}
\examples{

# tms_chromatogram <- function(
#   dat = pks,
#   comp = "15-HETE",
#   sid = "Sample1"
# )

}
