% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01params.R
\name{tms_format}
\alias{tms_format}
\title{Data Format (Targeted)}
\usage{
tms_format(dat, intype = "Skyline")
}
\arguments{
\item{dat}{Data object generated by msT_input function.}

\item{intype}{Exported data type (currently supports "Skyline").}
}
\value{
List containing formatted input data and metadata
for downstream analysis.
}
\description{
Formats input data from exported Skyline raw files. Must have
imported a valid sample list containing group information and from
msT_input and the exported sample_order if processing Skyline-based
targeted data.
}
\examples{

# msT_format(
#   dat = d1
# )

}
