% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01Tparams.R
\name{tms_input}
\alias{tms_input}
\title{Data Input (Targeted)}
\usage{
tms_input(dir1 = NULL)
}
\arguments{
\item{dir1}{Load all files from the specified directory if not NULL.}
}
\value{
List containing formatted input data and metadata
for downstream analysis.
}
\description{
Reads input data from exported Skyline raw files.
}
\examples{

# msT_input(
#   dir1 = "data/"
# )

}
