% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Tdataproc.r
\name{tms_integrate}
\alias{tms_integrate}
\title{Peak Integration (Targeted)}
\usage{
tms_integrate(
  dat,
  comp,
  sid,
  sidcol = "SampleID",
  intype = "Skyline",
  xcol = "RT_all",
  ycol = "Intensity",
  pkcol = "pkBoundary",
  phcol = "pkHeight",
  snr_type = "median",
  snr = 3
)
}
\arguments{
\item{dat}{Input data list.}

\item{comp}{Compound name to plot.}

\item{sid}{Name of a specific sample for peak integration.}

\item{sidcol}{Sample ID column.}

\item{intype}{Input data type: either "Skyline," a list of 2 vectors
containing values separated by commas (default) or "xy," which is
a list of two variables containing vectors of x and y coordinates.}

\item{xcol}{X column name.}

\item{ycol}{Y column name.}

\item{pkcol}{Peak boundary column, generated from tms_peakdetect.}

\item{phcol}{Peak height column.}

\item{snr_type}{Summary statistic for calculating signal to noise
ratio (either "median" or "mean").}

\item{snr}{Signal to noise ratio for filtering peaks.
Use 3 for LOD and 10 for LOQ.}
}
\value{
Integrated raw and normalized peak areas and heights.
}
\description{
Integrates detected peaks from a tms_peakdetect result.
}
\examples{

# d1 <- tms_integrate(
#   dat = pks,
#   comp = "15-HETE",
#   sid = "Stevens_879654_1_cal_8"
# )

}
