% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02dataproc.r
\name{tms_mzrt}
\alias{tms_mzrt}
\title{Retention Time Shift Summary (Targeted)}
\usage{
tms_mzrt(
  dat,
  ref,
  colcomp = "Name",
  colsid = "SampleID",
  colrtref = "ref_RT",
  colrtexp = "X"
)
}
\arguments{
\item{dat}{Input standard list.}

\item{ref}{Reference list containing retention times.}

\item{colcomp}{Variable containing standard names.}

\item{colsid}{Variable containing sample IDs.}

\item{colrtref}{Variable containing reference standard retention times.}

\item{colrtexp}{Variable containing measured retention times.}
}
\value{
mzrt scatter plot visualizing differences between
the expected and measured retention times of all reference
standards.
}
\description{
Mzrt scatterplot of internal standards highlighting
shift in expected vs. measured retention times.
}
\examples{

# d1 <- tms_mzrt(rt1, rt2)

}
