% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Tdataproc.r
\name{tms_peakdetect}
\alias{tms_peakdetect}
\title{Peak Detection (Targeted)}
\usage{
tms_peakdetect(
  dat,
  ref,
  intype = "Skyline",
  xcol = "RT_all",
  ycol = "Intensity",
  mcc = 8,
  comp,
  rttol = 0.2,
  spn = 7,
  lab = "SampleID",
  complab = "Name",
  rtcol = "exp_RT",
  snr_type = "median",
  snr = 3
)
}
\arguments{
\item{dat}{Input standard list.}

\item{ref}{Input mzrt list containing reference precursor m/z and
retention time columns.}

\item{intype}{Input data type: either "Skyline," a list of 2 vectors
containing values separated by commas (default) or "xy," which is
a list of two variables containing vectors of x and y coordinates.}

\item{xcol}{X column name.}

\item{ycol}{Y column name.}

\item{mcc}{Cores to use if not using Windows.}

\item{comp}{Compound name to plot.}

\item{rttol}{Retention time tolerance window (in minutes).}

\item{spn}{Number of points used for peak detection; must be an odd number.}

\item{lab}{Column containing sample labels.}

\item{complab}{Column containing compound names.}

\item{rtcol}{Retention time column.}

\item{snr_type}{Summary statistic to calculate signal-to-noise ratio
(either "median" [default] or "mean").}

\item{snr}{Signal to noise ratio for filtering peaks.
Use 3 for LOD and 10 for LOQ.}
}
\value{
mzrt scatter plot visualizing differences between
the expected and measured retention times of all reference
standards.
}
\description{
Calculates local minima and maxima within a specified
retention time window to determine peaks
}
\examples{

# d1 <- tms_peakdetect(
#   dat = d1[["merged"]],
#   ref = rtcor[["data"]],
#   comp = "15-HETE"
# )

}
