% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Tdataproc.r
\name{tms_quant}
\alias{tms_quant}
\title{Compound Quantification (Targeted)}
\usage{
tms_quant(dat, exdat, ref, intype = "Skyline", matd = "data", matm = "meta")
}
\arguments{
\item{dat}{List containing input data matrix containing samples as rows and
compounds as columns and a separate data frame containing metadata.}

\item{exdat}{Input data matrix of external calibration samples
provided in the same format as experimental samples
and a separate data frame containing metadata.}

\item{ref}{Input standard list with known concentrations and surrogate
standard associations.}

\item{intype}{Input data type: either "Skyline," a list of 2 vectors
containing values separated by commas (default) or "xy," which is
a list of two variables containing vectors of x and y coordinates.}

\item{comp}{Compound name to plot.}

\item{lab}{Column containing sample labels.}

\item{complab}{Column containing compound names.}

\item{rttol}{Retention time tolerance window (in minutes).}

\item{rtcol}{Retention time column.}

\item{snr_type}{Summary statistic to calculate signal-to-noise ratio
(either "median" [default] or "mean").}

\item{snr}{Signal to noise ratio for filtering peaks.
Use 3 for LOD and 10 for LOQ.}

\item{spn}{Number of points used for peak detection; must be an odd number.}

\item{mcc}{Cores to use if not using Windows.}
}
\value{
mzrt scatter plot visualizing differences between
the expected and measured retention times of all reference
standards.
}
\description{
Calculates molar concentrations in a targeted dataset
given a set of internal and surrogate standards. Uses
an area ratio approach to account for matrix effects.
}
\examples{

# d1 <- tms_peakdetect(
#   dat = d1[["merged"]],
#   ref = rtcor[["data"]],
#   comp = "15-HETE"
# )

}
