% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Tdataproc.r
\name{tms_rtcor}
\alias{tms_rtcor}
\title{Retention Time Correction (Targeted)}
\usage{
tms_rtcor(dat, ref, colrt = "ref_RT", colshift = "med_shift", type = "lm")
}
\arguments{
\item{dat}{Input standard list.}

\item{ref}{Reference list containing retention times.}

\item{colrt}{Retention time column.}

\item{colshift}{Variable containing measured retention time shifts.}

\item{type}{Regression type for predicting retention times ("lm" or "poly").}
}
\value{
mzrt scatter plot visualizing differences between
the expected and measured retention times of all reference
standards.
}
\description{
Applies retention time corrections based on observed
differences between expected vs. measured retention times
of reference standards in a targeted assay.
}
\examples{

# d1 <- tms_rtcor(ref = rtc, dat = mzlist)

}
