% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Tdataproc.r
\name{tms_rtshift}
\alias{tms_rtshift}
\title{Retention Time Shift (Targeted)}
\usage{
tms_rtshift(
  dat,
  mzrt,
  intype = "Skyline",
  xcol = "RT_all",
  ycol = "Intensity",
  mcc = 8,
  comp,
  rttol = 0.2,
  spn = 7,
  lab = "SampleID",
  sampid,
  snr_type = "median"
)
}
\arguments{
\item{dat}{Input standard list.}

\item{mzrt}{Input mzrt list containing reference precursor m/z and
retention time columns.}

\item{intype}{Input data type: either "Skyline," a list of 2 vectors
containing values separated by commas (default) or "xy," which is
a list of two variables containing vectors of x and y coordinates.}

\item{xcol}{X column name.}

\item{ycol}{Y column name.}

\item{mcc}{Cores to use if not using Windows.}

\item{comp}{Compound name to plot.}

\item{rttol}{Retention time tolerance window (in minutes).}

\item{spn}{Number of points used for peak detection; must be an odd number.}

\item{lab}{Column containing sample labels.}

\item{sampid}{Name of a specific sample for peak integration.}

\item{snr_type}{Summary statistic to calculate signal-to-noise ratio
(either "median" [default] or "mean").}
}
\value{
A data frame containing the calculated difference between
the expected and measured retention times for a list of reference
standards.
}
\description{
Calculates retention time shifts of standards
used for data acquisition.
}
\examples{

# msT_integrate <- function(
#   dat = d1[["merged"]][d1[["merged"]][["Group"]] == "Std_curve", ],
#   comp = "15-HETE",
#   ref = "Std_curve",
#   grp = "Std_curve",
#   sampid = "Stevens_879654_1_cal_8"
# )

}
