% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02datanorm.r
\name{ms_data_norm}
\alias{ms_data_norm}
\title{Data Normalization}
\usage{
ms_data_norm(ld1, mtd, parl = FALSE, core_perc = NULL)
}
\arguments{
\item{ld1}{List data object generated from ms_input().}

\item{mtd}{Normalization method (either "none" or "tic").}

\item{parl}{(optional) Logical indicating if normalization
should be performed in parallel.}

\item{core_perc}{(optional) Percentage of cores to use for parallel
normalization.}
}
\value{
A list containing normalized compound intensities and metadata.
}
\description{
Includes multiple methods for data normalization. Currently supported methods
are total ion count (TIC) normalization.
Planned methods include iSTD, SERRF, and LOESS.
}
\examples{

# d_norm <- ms_data_norm(
#   dm = d_norm[["Data"]],
#   mtd = "tic"
# )

}
