% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03datanorm.r
\name{ms_plot_tic}
\alias{ms_plot_tic}
\title{Normalization QC}
\usage{
ms_plot_tic(df, var_x, var_g = "Group")
}
\arguments{
\item{df}{input data from list data object.}

\item{var_x}{X-axis variable (usually sample ID).}

\item{var_g}{Grouping variable.}
}
\value{
A scatter plot visualizing overall sample intensities.
}
\description{
Scatter plot visualization to evaluate normalization performance.
}
\examples{

# ptic <- msi_plot_tic(
#   df = ld[["meta"]],
#   var_x = "sampleID",
#   var_g = "Group"
# )

}
