% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09visualize.R
\name{ms_plot_vio}
\alias{ms_plot_vio}
\title{Violin Plot}
\usage{
ms_plot_vio(
  df1,
  xvar,
  yvar,
  clvar = "Group",
  plab = "nM / ng per million cells"
)
}
\arguments{
\item{df1}{Input data frame.}

\item{xvar}{X-axis variable.}

\item{yvar}{Y-axis variable.}

\item{clvar}{Group variable.}

\item{plab}{Plot y-axis label.}
}
\value{
A violin plot for the chosen treatment comparisons.
}
\description{
Generates a violin plot from an input data frame.
}
\examples{

# ms_plot_violin(
#   df1 = dox,
#   xvar = "rt",
#   yvar = "mz"
# )

}
