% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05fc.R
\name{ms_stat_fc}
\alias{ms_stat_fc}
\title{Group-wise Fold Change Function}
\usage{
ms_stat_fc(
  mat1,
  mat_type = "norm",
  md,
  md_var,
  an = NULL,
  fc_class = FALSE,
  grp_class = NULL,
  lipid_fc = FALSE,
  comp_grp = TRUE
)
}
\arguments{
\item{mat1}{Data matrix assigned by ms_input().}

\item{mat_type}{Input data type, either "norm" or "scaled." If
"scaled," fold changes are calculated as the difference between
two group averages rather than by dividing normalized compound
average intensities per group.}

\item{md}{Dataframe containing study metadata.}

\item{md_var}{Vector of metadata variables to calculate fold
changes for.}

\item{an}{Dataframe containing annotation metadata.}

\item{fc_class}{Logical indicating whether class-wide
fold changes should be calculated. Requires a valid
annotation list contained within ms_input() result.}

\item{grp_class}{Name of the class annotation column
to use. Ignored when fc_class is set to FALSE.}

\item{lipid_fc}{(Lipidomics Only) Logical indicating
whether related lipid class fold changes should be calculated.
If TRUE, calculates lipid class averages and calculates related class
ratios prior to treatment group comparisons.}

\item{comp_grp}{Only applicable if lipid_fc is TRUE; Should treatment
group fold changes be calculated? If FALSE, returns lipid class ratios
without aggregating and comparing by treatment group.}
}
\value{
A data frame containing the group-wise
fold changes for each compound.
}
\description{
Calculates the log2 fold changes for all
possible combinations of selected metadata variables.
}
\examples{

# ms_stat_fc(
#   mat1 = d[["data"]][["imputed"]],
#   md = d[["data"]][["meta"]],
#   md_var = c("Organ", "Sex", "Dose"),
#   an = d[["data"]][["anno"]],
#   fc_class = TRUE,
#   grp_class = "Subclass"
# )

}
