% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00lib.R
\name{ms_theme}
\alias{ms_theme}
\title{Generic Plot Theme}
\usage{
ms_theme(leg = c(0.95, 0.95))
}
\arguments{
\item{leg}{Plot legend coordinates.}
}
\value{
ggplot2 theme parameters to replace default plot theme.
}
\description{
General plotting theme.
}
\examples{

# ms_theme()

}
